# docs/swagger.yaml
basePath: /v1
definitions:
  clients.SalesforceAndamento:
    properties:
      dataInsercao:
        type: string
      descricao:
        type: string
      evento:
        type: string
      status:
        type: string
      tipoEvento:
        type: string
    type: object
  clients.SalesforceAnonimizacaoStatus:
    properties:
      cpf:
        type: string
      dataConclusao:
        type: string
      dataSolicitacao:
        type: string
      numeroSolicitacao:
        type: string
      status:
        type: string
    type: object
  clients.SalesforceCanOpen:
    properties:
      elogio:
        $ref: '#/definitions/clients.SalesforceCanOpenAction'
      reclamacao:
        $ref: '#/definitions/clients.SalesforceCanOpenAction'
      sugestao:
        $ref: '#/definitions/clients.SalesforceCanOpenAction'
    type: object
  clients.SalesforceCanOpenAction:
    properties:
      available:
        type: boolean
      categoryId:
        type: string
    type: object
  clients.SalesforceChamadosList:
    properties:
      ouvidoriaWindow:
        $ref: '#/definitions/clients.SalesforceOuvidoriaWindow'
      protocolos:
        items:
          $ref: '#/definitions/clients.SalesforceProtocolo'
        type: array
    type: object
  clients.SalesforceCidadao:
    properties:
      accountId:
        type: string
      canalOrigem:
        type: string
      canalUltimaModificacao:
        type: string
      complemento:
        type: string
      consentimento:
        items:
          $ref: '#/definitions/clients.SalesforceConsentimento'
        type: array
      cpf:
        type: string
      dataNascimento:
        type: string
      deficiencia:
        type: string
      email:
        type: string
      endereco:
        $ref: '#/definitions/clients.SalesforceEndereco'
      escolaridade:
        type: string
      genero:
        type: string
      idioma:
        items:
          type: string
        type: array
      isTourist:
        type: boolean
      nacionalidade:
        type: string
      nome:
        type: string
      nomeExibicao:
        type: string
      nomeSocial:
        type: string
      passaporte:
        type: string
      raca:
        type: string
      rendaFamiliar:
        type: string
      telefone1:
        type: string
      telefone2:
        type: string
      telefone3:
        type: string
      telefoneInternacional:
        type: string
      telefonePrincipal:
        type: string
      tipoTelefone1:
        type: string
      tipoTelefone2:
        type: string
      tipoTelefone3:
        type: string
    type: object
  clients.SalesforceConsentimento:
    properties:
      acao:
        description: 'legacy: optin | optout'
        type: string
      canal:
        type: string
      categoria:
        type: string
      data:
        type: string
      finalidade:
        type: string
      motivo:
        description: required when status=OUT / acao=optout
        type: string
      status:
        description: IN | OUT
        type: string
    type: object
  clients.SalesforceConsentimentoPatchRequest:
    properties:
      acao:
        type: string
      categoria:
        type: string
      motivo:
        type: string
      origem:
        type: string
    type: object
  clients.SalesforceEndereco:
    properties:
      bairro:
        type: string
      cep:
        type: string
      cidade:
        type: string
      complemento:
        type: string
      estado:
        type: string
      logradouro:
        type: string
      pais:
        type: string
    type: object
  clients.SalesforceEnderecoChamado:
    properties:
      bairro:
        type: string
      cep:
        type: string
      complemento:
        type: string
      coordenadaX:
        type: number
      coordenadaY:
        type: number
      logradouro:
        type: string
      numero:
        type: string
      pontoReferencia:
        type: string
      tipoEndereco:
        type: string
    type: object
  clients.SalesforceErrorBody:
    properties:
      errors:
        items:
          $ref: '#/definitions/clients.SalesforceFieldError'
        type: array
    type: object
  clients.SalesforceExportacao:
    properties:
      cpf:
        type: string
      dadosPessoais:
        additionalProperties: true
        type: object
      dataExportacao:
        type: string
      relacionados:
        additionalProperties: true
        type: object
    type: object
  clients.SalesforceFieldError:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
  clients.SalesforceOrdemServico:
    properties:
      andamentos:
        items:
          $ref: '#/definitions/clients.SalesforceAndamento'
        type: array
      bairro:
        type: string
      canOpen:
        $ref: '#/definitions/clients.SalesforceCanOpen'
      categoria:
        type: string
      codigoOs:
        type: string
      dataAbertura:
        type: string
      dataFechamento:
        type: string
      dataTermino:
        type: string
      descricao:
        type: string
      endereco:
        $ref: '#/definitions/clients.SalesforceEnderecoChamado'
      id:
        type: string
      idCatalogoServico:
        type: string
      isClosed:
        type: boolean
      isOverdue:
        type: boolean
      orgaoResponsavel:
        $ref: '#/definitions/clients.SalesforceOrgao'
      orgaoResponsavelPai:
        $ref: '#/definitions/clients.SalesforceOrgao'
      ouvidoriaUrl:
        type: string
      ouvidoriaWindow:
        $ref: '#/definitions/clients.SalesforceOuvidoriaWindow'
      previsaoSLA:
        type: string
      prioridade:
        type: string
      servico:
        type: string
      status:
        type: string
      subtema:
        type: string
      tema:
        type: string
    type: object
  clients.SalesforceOrgao:
    properties:
      id:
        type: string
      nome:
        type: string
      sigla:
        type: string
    type: object
  clients.SalesforceOuvidoriaWindow:
    properties:
      days:
        type: integer
      label:
        type: string
    type: object
  clients.SalesforceProtocolo:
    properties:
      categoria:
        type: string
      dataAbertura:
        type: string
      dataFechamento:
        type: string
      dataUltimaAtualizacao:
        type: string
      ordens_de_servico:
        items:
          $ref: '#/definitions/clients.SalesforceOrdemServico'
        type: array
      origem:
        type: string
      protocolo:
        type: string
      status:
        type: string
      subcategoria:
        type: string
    type: object
  handlers.AuthValidateResponse:
    properties:
      action:
        enum:
        - matched
        - updated
        - created
        example: matched
        type: string
      cidadao:
        $ref: '#/definitions/clients.SalesforceCidadao'
    type: object
...
  handlers.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: suporte@rio.rj.gov.br
    name: Suporte RMI
    url: http://www.rio.rj.gov.br
  description: API para gerenciamento de dados de cidadãos do Rio de Janeiro, incluindo
    autodeclaração de informações e verificação de contato.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: API RMI
  version: "1.0"
paths:
...
  /auth/validate:
    get:
      description: 'Orquestra o Person Account no Salesforce: GET → match / PATCH
        silencioso de email / POST create em 404 (contaOrigem Portal Pref.Rio). Usa
        CPF, nome e email do JWT. Não persiste dados do SF no Mongo do RMI.'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.AuthValidateResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "502":
          description: Falha ao chamar Salesforce
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "503":
          description: SALESFORCE_BASE_URL não configurado
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Sincronizar Person Account no login
      tags:
      - salesforce
...
  /salesforce/anonimizacao/{numeroSolicitacao}:
    get:
      description: Proxy para GET /api/private/anonimizacao/{numeroSolicitacao}.
      parameters:
      - description: Número da solicitação
        in: path
        name: numeroSolicitacao
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/clients.SalesforceAnonimizacaoStatus'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/clients.SalesforceErrorBody'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Consultar status de anonimização
      tags:
      - salesforce
  /salesforce/chamados:
    get:
      description: Proxy para GET /api/private/chamados usando o JWT do usuário (Community
        User).
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/clients.SalesforceChamadosList'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Listar chamados do cidadão
      tags:
      - salesforce
  /salesforce/chamados/{protocolo}:
    get:
      description: Proxy para GET /api/private/chamados/{protocolo}.
      parameters:
      - description: Protocolo do chamado
        in: path
        name: protocolo
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/clients.SalesforceProtocolo'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/clients.SalesforceErrorBody'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Detalhar chamado por protocolo
      tags:
      - salesforce
  /salesforce/cidadao/{cpf}/anonimizar:
    post:
      description: Proxy para POST /api/private/cidadao/{cpf}/anonimizar (assíncrono).
        202 enfileirado; 409 se já houver solicitação aberta.
      parameters:
      - description: CPF do cidadão
        in: path
        name: cpf
        required: true
        type: string
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/clients.SalesforceAnonimizacaoStatus'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "409":
          description: Solicitação já em aberto
          schema:
            $ref: '#/definitions/clients.SalesforceAnonimizacaoStatus'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Solicitar anonimização no Salesforce
      tags:
      - salesforce
  /salesforce/cidadao/{cpf}/consentimento:
    patch:
      consumes:
      - application/json
      description: Proxy para PATCH /api/private/cidadao/{cpf}/consentimento. Encaminha
        o JWT do usuário. categoria deve ser valor de picklist SF (ex. PREF_Lembrete_Pagamento).
        motivo obrigatório quando acao=optout.
      parameters:
      - description: CPF do cidadão
        in: path
        name: cpf
        required: true
        type: string
      - description: Consentimento
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/clients.SalesforceConsentimentoPatchRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Consentimento atualizado (sem corpo)
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: CPF não pertence ao usuário autenticado
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/clients.SalesforceErrorBody'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/clients.SalesforceErrorBody'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Atualizar consentimento no Salesforce
      tags:
      - salesforce
  /salesforce/cidadao/{cpf}/exportar:
    get:
      description: Proxy síncrono para GET /api/private/cidadao/{cpf}/exportar.
      parameters:
      - description: CPF do cidadão
        in: path
        name: cpf
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/clients.SalesforceExportacao'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/clients.SalesforceErrorBody'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Exportar dados pessoais (LGPD) do Salesforce
      tags:
      - salesforce
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: 'Tipo: Bearer token. Exemplo: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...